PRLT-939: Add Drizzle migration system for SQLite schema upgrades
Add a tracked migration system so schema changes between versions
upgrade cleanly instead of relying on scattered inline ALTER TABLE
statements.

- Extract CREATE_TABLES_SQL to workspace-schema.ts to avoid circular imports
- Add migrator.ts with runDrizzleMigrations() that tracks applied
  migrations in a prlt_migrations table
- Add migrations/0001_baseline.ts as the initial schema snapshot
  (all CREATE TABLE IF NOT EXISTS, safe on new and existing DBs)
- Integrate into openWorkspaceDatabase() and createWorkspaceDatabase()
- Existing inline ALTER TABLE migrations kept for backward compat
